third_party/blink/web_tests/external/wpt/css/css-display/parsing/tentative/display-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property display value 'grid-lanes'
  assert_true: 'grid-lanes' is a supported value for display. expected true got false
[FAIL] Property display value 'inline-grid-lanes'
  assert_true: 'inline-grid-lanes' is a supported value for display. expected true got false
[FAIL] position absolute affects computed display
  assert_equals: inline-masonry -> masonry expected "grid-lanes" but got "block"
[FAIL] position fixed affects computed display
  assert_equals: inline-masonry -> masonry expected "grid-lanes" but got "block"
[FAIL] float left affects computed display
  assert_equals: inline-masonry -> masonry expected "grid-lanes" but got "block"
[FAIL] float right affects computed display
  assert_equals: inline-masonry -> masonry expected "grid-lanes" but got "block"
Harness: the test ran to completion.
